+++ specifications/src/main/java/org/dcsa/conformance/specifications/an/v100/EnumBase.java
@@ -1,6 +1,0 @@
-package org.dcsa.conformance.specifications.an.v100;
-
-public interface EnumBase {
-  String name();
-  String getValueDescription();
-}