+++ specifications/src/main/java/org/dcsa/conformance/specifications/generator/StringPatterns.java
@@ -1,0 +1,5 @@
+package org.dcsa.conformance.specifications.generator;
+
+public class StringPatterns {
+  public static final String TRIMMED_NON_EMPTY_STRING = "^\\S(?:.*\\S)?$";
+}